conversation Db handlers implemented

+++ models/conversationStorage.ts
@@ -8,11 +8,17 @@ import {
     updateConversationDB,
     deleteConversationDB
 } from "./conversationStorageDB"
+import { newConversationName } from "../helpers/conversations"
 
 const conversationsDirectory = "./logs/conversations/";
 
-export async function deleteConversation(conversationName: string): Promise<void> {
-    const File = path.join(conversationsDirectory, conversationName);
+export async function deleteConversation(conversationName: string, userId: number): Promise<void> {
+    const conversation = await getConversationDB(conversationName, userId)
+    if (conversation == null || !conversation.uniqueId) throw Error("Conversation not found in db")
+
+    await deleteConversationDB(conversation.uniqueId)
+
+    const File = path.join(conversationsDirectory, conversation.uniqueId + '.json');
     try {
         fs.unlinkSync(File);
     } catch (error) {
@@ -24,26 +30,13 @@ export async function deleteConversation(conversationName: string): Promise<void
 export async function createConversation(conversationName: string, userId: number): Promise<{ fileName: string; content: any }> {
 
     if (conversationName == null || conversationName == "") {
-        const date = new Date();
-        conversationName = date
-            .toDateString()
-            .replace(/\s/g, "") +
-            "-" +
-            date
-                .toLocaleTimeString()
-                .replace(":", "-")
-                .replace(/\s/g, "")
-                .replace(":", "-")
-                .replace(/\s/g, "");
+        conversationName = newConversationName()
     }
 
     try {
 
         const newConversation = await createConversationDB(userId);
-
-        if (!newConversation.uniqueId) {
-            throw Error("Error creating conversation file in db")
-        }
+        if (!newConversation.uniqueId) throw Error("Error creating conversation file in db")
 
         if (!fs.existsSync(conversationsDirectory)) {
             fs.mkdirSync(conversationsDirectory, { recursive: true });
@@ -62,16 +55,18 @@ export async function createConversation(conversationName: string, userId: numbe
     }
 }
 
-export async function getConversation(conversationId: string): Promise<any> {
+export async function getConversation(conversationId: string, userId: number): Promise<any> {
     try {
-        const filePath = path.join(conversationsDirectory, conversationId);
+        const conversation = await getConversationDB(conversationId, userId)
+        if (conversation == null || !conversation.uniqueId) throw Error("Conversation not found in db")
+
+        const filePath = path.join(conversationsDirectory, conversation.uniqueId + '.json');
 
         if (fs.existsSync(filePath)) {
             const rawData = fs.readFileSync(filePath, "utf8");
             return JSON.parse(rawData);
         } else {
             console.log('no file');
-
             return undefined;
         }
     } catch (error) {
@@ -80,24 +75,31 @@ export async function getConversation(conversationId: string): Promise<any> {
     }
 }
 
-export async function setConversation(conversationId: string, conversation: any): Promise<void> {
+export async function setConversation(conversationId: string, conversation: any, userId: number): Promise<void> {
     try {
-        const filePath = path.join(conversationsDirectory, conversationId);
+        const db_conversation = await getConversationDB(conversationId, userId)
+        if (db_conversation == null || !db_conversation.uniqueId) throw Error("Conversation not found in db")
+
+        await updateConversationDB(db_conversation.uniqueId, new Date())
+
+        const filePath = path.join(conversationsDirectory, db_conversation.uniqueId + '.json');
         fs.writeFileSync(filePath, JSON.stringify(conversation, null, 2));
     } catch (error) {
         console.error("Error writing conversation file:", error);
     }
 }
 
-export async function getAllConversations(): Promise<{ fileName: string; docname: string }[]> {
+export async function getAllConversations(userId: number): Promise<{ fileName: string; docname: string }[]> {
+
     try {
-        const fileNames = fs.readdirSync(conversationsDirectory);
-        const conversations = fileNames.map((fileName) => {
-            const filePath = path.join(conversationsDirectory, fileName);
+        const fileNames = await getAllConversationsDB(userId)
+
+        const conversations: any = fileNames.map((fileName) => {
+            const filePath = path.join(conversationsDirectory, fileName.uniqueId + '.json');
             const rawData = fs.readFileSync(filePath, "utf8");
             const content = JSON.parse(rawData);
             return {
-                fileName: fileName,
+                fileName: fileName.uniqueId,
                 docname: content[0].fileName,
             };
         });
@@ -108,18 +110,24 @@ export async function getAllConversations(): Promise<{ fileName: string; docname
     }
 }
 
-export async function changeName(conversationName: string, newName: string): Promise<void> {
+export async function changeName(conversationName: string, newName: string, userId: number): Promise<void> {
+
+    const db_conversation = await getConversationDB(conversationName, userId)
+    if (db_conversation == null || !db_conversation.uniqueId) throw Error("Conversation not found in db")
+
+    await updateConversationDB(db_conversation.uniqueId, new Date())
 
     try {
-        const filePath = path.join(conversationsDirectory, conversationName);
+        const filePath = path.join(conversationsDirectory, conversationName + '.json');
 
         if (fs.existsSync(filePath)) {
             const rawData = fs.readFileSync(filePath, "utf8");
             let jsonData = JSON.parse(rawData);
             jsonData[0].fileName = newName;
-            setConversation(conversationName, jsonData);
+            setConversation(conversationName, jsonData, userId);
         } else {
-            console.log("File not exist");
+            console.log("File not exist")
+            throw Error("File not exist")
         }
     } catch (error) {
         console.log("Error reading conversation file", error);
